requirements.txt
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
pillow==11.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pycparser==2.22
PyJWT==2.10.0
pyOpenSSL==24.3.0
pyotp==2.9.0
qrcode==8.0
requests==2.32.3
service-identity==24.2.0
sqlparse==0.5.2
tomli==2.2.1
twisted==24.11.0
txaio==23.1.1
typing-extensions==4.12.2
urllib3==2.2.3
zope.interface==7.2
